Script to check deploy status Restart the nginx service from nginx container and check wether the overlay network working correctly if not remove the stack and redeploy it

+++ hamon/healthcheck.sh
@@ -1,15 +1,31 @@
 #!/bin/bash
 LOG_FILE=deploy.log
 
-# Remove existing story stack
-echo "Going to remove existing stack" | tee -a $LOG_FILE   
-sudo docker stack rm story
+# Check whether the nginx running correctly
 
-# Wait for a while for remove stack
-echo "Removing....wait...a..while"
-sleep(60)
+# Get the container id of nginx
+while [ -z $CONTAINER_ID ]; do
+    CONTAINER_ID=$(sudo docker ps -qf name=story_sw-nginx)
+    sleep 10
+done
+
+#Try to restart nginx service
+RETURN_MSG=$(sudo docker exec -t  $CONTAINER_ID service nginx restart)
+
+#check wether have DNS issue if have redeploy the stack
+if [[ "$RETURN_MSG" =~ "host not found in upstream /"sw-app/" in /etc/nginx/sites" ]]; then
+        
+    # Remove existing story stack
+    echo "Going to remove existing stack" | tee -a $LOG_FILE   
+    sudo docker stack rm story
+
+    # Wait for a while for remove stack
+    echo "Removing....wait...a..while"
+    sleep 60
 
 
-# Redeploy the stack
-echo "Going to Redeploy" | tee -a $LOG_FILE   
-sudo docker stack deploy -c docker-compose.yml story
+    # Redeploy the stack
+    echo "Going to Redeploy" | tee -a $LOG_FILE   
+    sudo docker stack deploy -c docker-compose.yml story
+    sleep 60
+fi
